server/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.9
aiosqlite==0.20.0
pydantic==2.9.2
pydantic-settings==2.5.2
Pillow==10.3.0
httpx==0.27.0
